move function to corrent file
